backend/requirements.txt
altair==5.4.1
altgraph==0.17.4
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==44.0.0
et-xmlfile==1.1.0
firebase-admin==6.6.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.20.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
msgpack==1.1.0
mysql-connector-python==9.2.0
narwhals==1.8.1
numpy==2.1.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pillow==10.4.0
proto-plus==1.25.0
protobuf==5.28.2
py-xbrl==2.2.12
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
soupsieve==2.6
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
XlsxWriter==3.2.0
fastapi>=0.68.0
uvicorn>=0.15.0
uvicorn[standard]
requests>=2.26.0
pydantic
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
textblob>=0.15.3
OpenAI
email-validator
pandas
nltk
asyncpg
aiohttp>=3.9.0
mcp>=1.0.0
